Usando a Target class para se relacionar com as futuras classes contendo diferentes tipos de alvos, criando um array contendo as mesmas, para facilitar a renderização e a verificação da lógica do jogo

+++ EP2/src/dev/ep2/battleship/entities/Entity.java
@@ -6,7 +6,7 @@ public abstract class Entity {
 
 	protected float x,y; // float for smooth movement
 	
-	public Entity(float x, float y) {
+	public Entity(int x, int y) {
 		
 		this.x = x;
 		this.y = y;
